Extract app-server event compaction helpers
